ci: test packages
