+++ .github/workflows/build.yml
@@ -31,3 +31,5 @@ jobs:
         run: pnpm lint
       - name: Run build
         run: pnpm build
+      - name: Run test
+        run: pnpm test
