+++ packages/protocol/src/stream.test.ts
@@ -1,0 +1,9 @@
+describe('StreamMessagesStream', () => {
+  it('produces data', async () => {})
+
+  describe('on error', () => {
+    it('produces an error by default', () => {})
+    it('reconnects', () => {})
+    it('produces an error after trying to reconnect', () => {})
+  })
+})
