+++ packages/starknet/jest.config.ts
@@ -1,0 +1,8 @@
+import { Config } from 'jest'
+
+const config: Config = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+}
+
+export default config
