+++ packages/starknet/tsconfig.json
@@ -4,6 +4,8 @@
     "outDir": "dist",
     "declarationDir": "dist",
     "noEmit": false,
-    "rootDir": "src"
-  }
+    "rootDir": "src",
+    "types": ["jest", "node"]
+  },
+  "include": ["src/"]
 }
